rpi: Generate a unique host name based on eth0 MAC address
The generated host name is 'rpi' followed by the 6 last digits of the MAC
address (i.e. without the OUI, which is constant for all Raspberry Pis).

This avoids NetBIOS name conflicts if several boards are connected to the same
network.

This operation is triggered by 'ifup eth0', and not only by the system startup,
because the Ethernet support is based on a USB chip, which may become available
too late during startup. The MAC address from the kernel command line could also
have been used, but the solution chosen here has the advantage of not relying on
the way the MAC address is passed to the Ethernet controller.

nmbd is restarted upon 'ifup eth0' in order to take the new host name into
account for the NetBIOS server.

If no MAC address is found, then the default 'rpi' host name is used.

Signed-off-by: Benoît Thébaudeau <benoit.thebaudeau@advansee.com>